fix: vault_credential limited-networking and mcp_server_url create bugs
Add live acceptance tests (TestAcc) for anthropic_vault and
anthropic_vault_credential, which surfaced two create-path bugs that
plan-only and unit tests could not catch:

- environment_variable + limited networking failed at create with
  "networking.type: Field required": the SDK's
  BetaManagedAgentsCredentialNetworkingParamsOfLimited helper omits the
  "type" discriminator. Construct the limited variant explicitly with
  Type "limited".
- mcp_server_url with a bare host (no path) produced "inconsistent result
  after apply": the API normalizes "https://host" to "https://host/".
  Since mcp_server_url is RequiresReplace (not Computed), preserve the
  equivalent user value (modulo trailing slash) via mcpServerURLForState
  to avoid the inconsistency and a perpetual replace diff.

Vaults are a standard-API resource scoped to the test key's workspace, so
(unlike the admin-API resources blocked by #58) they support full CRUD
acceptance tests; credentials are not validated until session runtime, so
tests use placeholder secrets and unreachable MCP URLs. Flip the
vault_credential native test to apply now that creation works end-to-end.

Rename the internal unit-test files to the _mapping_test.go convention so
the plain _resource_test.go names hold the external acceptance tests.
