+++ CLAUDE.md
@@ -93,6 +93,7 @@ internal/
 - `internal/services/workspaces/workspacetest.go` defines workspaces-specific shared helpers (`workspaceFixture`, `pageData`, `fetchAllPages`, plus a thin `newTestAdminClient` that delegates to `admintest.NewClient`); reuse them in any new workspaces unit test instead of duplicating pagination loops
 - **Admin API acceptance tests for read-only data sources:** target the `terraform-tests` workspace via `acctest.TerraformTestsWorkspaceID` and gate on `acctest.PreCheckAdmin`. These are **smoke** tests (the live workspace's data isn't deterministic, so assert attribute presence like `members.#`, not specific values); they **complement, not replace**, the httptest-based unit tests that deterministically cover pagination, query-param/filter construction, mapping edge cases, and 404 handling. Covered so far: `workspace`, `workspaces`, `workspace_members`, `workspace_rate_limits`, `organization`, `api_keys`. Resource tests (create/update/delete) on the Admin API remain blocked by [#58](https://github.com/ippontech/terraform-provider-anthropic/issues/58) because we do not yet have a dedicated test organization, so do not create new resources via Admin API in tests.
 - Admin API **Terraform native tests**: use `command = plan` (not the default `apply`) until a test org is available; this validates schema without making live API calls. Read-only data source native tests may run against the `terraform-tests` workspace ID above.
+- **Standard-API resources support full CRUD acceptance tests.** The #58 blocker is Admin-API-only. Vaults and vault credentials use the standard `ANTHROPIC_API_KEY` (gate on `acctest.PreCheck`) and are workspace-scoped to that key's workspace — there is no `workspace_id` create param, so they land wherever the key is scoped (point the test key at `terraform-tests`). Vaults are billed only at runtime, so create/destroy is free. **Vault credentials are not validated until session runtime**, so acceptance tests create them with placeholder secrets and unreachable `mcp_server_url`s; the full CRUD lifecycle runs without ever starting a session. Their native test (`tests/vault_credential.tftest.hcl`) therefore uses `command = apply`, not plan. `archive_on_destroy` tests for a credential must set `archive_on_destroy = true` on the parent vault too (a hard-deleted vault cascade-deletes the credential, leaving nothing to assert on).
 - Unit tests: no special env vars needed
 - Acceptance tests: use `resource.Test(t, resource.TestCase{...})` with `TF_ACC=1`
