+++ internal/services/vaults/vault_resource_test.go
@@ -1,152 +1,216 @@
 // Copyright (c) Ippon
 // SPDX-License-Identifier: MPL-2.0
 
-package vaults
+package vaults_test
 
 import (
+	"context"
+	"fmt"
+	"os"
 	"testing"
-	"time"
+
+	acctest "github.com/ippontech/terraform-provider-anthropic/internal/acctest"
 
 	"github.com/anthropics/anthropic-sdk-go"
-	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/anthropics/anthropic-sdk-go/option"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
 )
 
-func TestMapVaultResponseToState_BasicFields(t *testing.T) {
-	createdAt := time.Date(2024, 1, 15, 10, 0, 0, 0, time.UTC)
-	updatedAt := time.Date(2024, 1, 15, 11, 0, 0, 0, time.UTC)
+// Vaults are a standard-API (non-admin) resource scoped to the workspace of the
+// ANTHROPIC_API_KEY used by the test, so they support full create/read/update/
+// delete acceptance tests (unlike the admin-API resources blocked by #58).
+// Vaults are billed only at runtime, so creating and destroying them in tests is
+// free; CheckDestroy guarantees no dangling resources remain.
 
-	vault := &anthropic.BetaManagedAgentsVault{
-		ID:          "vlt_01ABC",
-		DisplayName: "my-vault",
-		Type:        anthropic.BetaManagedAgentsVaultTypeVault,
-		CreatedAt:   createdAt,
-		UpdatedAt:   updatedAt,
-		Metadata:    nil,
+func testAccCheckVaultDestroyed(s *terraform.State) error {
+	client := anthropic.NewClient(option.WithAPIKey(os.Getenv("ANTHROPIC_API_KEY")))
+	for _, rs := range s.RootModule().Resources {
+		if rs.Type != "anthropic_vault" {
+			continue
+		}
+		_, err := client.Beta.Vaults.Get(context.Background(), rs.Primary.ID, anthropic.BetaVaultGetParams{})
+		if err != nil {
+			// Resource not found — destroyed successfully.
+			return nil
+		}
+		return fmt.Errorf("vault %s still exists", rs.Primary.ID)
 	}
-
-	var data VaultResourceModel
-	diags := mapVaultResponseToState(vault, &data)
-	if diags.HasError() {
-		t.Fatalf("unexpected diagnostics: %v", diags)
-	}
-
-	if data.ID.ValueString() != "vlt_01ABC" {
-		t.Errorf("expected ID vlt_01ABC, got %s", data.ID.ValueString())
-	}
-	if data.DisplayName.ValueString() != "my-vault" {
-		t.Errorf("expected DisplayName my-vault, got %s", data.DisplayName.ValueString())
-	}
-	if data.Type.ValueString() != "vault" {
-		t.Errorf("expected Type vault, got %s", data.Type.ValueString())
-	}
-	if data.CreatedAt.ValueString() != "2024-01-15T10:00:00Z" {
-		t.Errorf("expected CreatedAt 2024-01-15T10:00:00Z, got %s", data.CreatedAt.ValueString())
-	}
-	if data.UpdatedAt.ValueString() != "2024-01-15T11:00:00Z" {
-		t.Errorf("expected UpdatedAt 2024-01-15T11:00:00Z, got %s", data.UpdatedAt.ValueString())
-	}
+	return nil
 }
 
-func TestMapVaultResponseToState_ArchivedAtZero(t *testing.T) {
-	vault := &anthropic.BetaManagedAgentsVault{
-		ID:          "vlt_02DEF",
-		DisplayName: "unarchived",
-		Type:        anthropic.BetaManagedAgentsVaultTypeVault,
-		CreatedAt:   time.Now(),
-		UpdatedAt:   time.Now(),
-		// ArchivedAt is zero value (not archived)
+// testAccCheckVaultArchivedAndCleanup verifies the vault was archived (not
+// hard-deleted) and then permanently deletes it to avoid dangling resources.
+func testAccCheckVaultArchivedAndCleanup(s *terraform.State) error {
+	client := anthropic.NewClient(option.WithAPIKey(os.Getenv("ANTHROPIC_API_KEY")))
+	for _, rs := range s.RootModule().Resources {
+		if rs.Type != "anthropic_vault" {
+			continue
+		}
+		vault, err := client.Beta.Vaults.Get(context.Background(), rs.Primary.ID, anthropic.BetaVaultGetParams{})
+		if err != nil {
+			return fmt.Errorf("vault %s not found after archive destroy: %w", rs.Primary.ID, err)
+		}
+		if vault.ArchivedAt.IsZero() {
+			return fmt.Errorf("vault %s was not archived on destroy", rs.Primary.ID)
+		}
+		// Hard-delete the archived vault so it doesn't accumulate in the workspace.
+		if _, err := client.Beta.Vaults.Delete(context.Background(), rs.Primary.ID, anthropic.BetaVaultDeleteParams{}); err != nil {
+			return fmt.Errorf("cleanup: unable to delete archived vault %s: %w", rs.Primary.ID, err)
+		}
 	}
-
-	var data VaultResourceModel
-	diags := mapVaultResponseToState(vault, &data)
-	if diags.HasError() {
-		t.Fatalf("unexpected diagnostics: %v", diags)
-	}
-
-	if !data.ArchivedAt.IsNull() {
-		t.Errorf("expected ArchivedAt to be null for non-archived vault, got %s", data.ArchivedAt.ValueString())
-	}
+	return nil
 }
 
-func TestMapVaultResponseToState_ArchivedAtNonZero(t *testing.T) {
-	archivedAt := time.Date(2024, 6, 1, 12, 0, 0, 0, time.UTC)
-	vault := &anthropic.BetaManagedAgentsVault{
-		ID:          "vlt_03GHI",
-		DisplayName: "archived-vault",
-		Type:        anthropic.BetaManagedAgentsVaultTypeVault,
-		CreatedAt:   time.Now(),
-		UpdatedAt:   time.Now(),
-		ArchivedAt:  archivedAt,
-	}
-
-	var data VaultResourceModel
-	diags := mapVaultResponseToState(vault, &data)
-	if diags.HasError() {
-		t.Fatalf("unexpected diagnostics: %v", diags)
-	}
-
-	if data.ArchivedAt.IsNull() {
-		t.Error("expected ArchivedAt to be non-null for archived vault")
-	}
-	if data.ArchivedAt.ValueString() != "2024-06-01T12:00:00Z" {
-		t.Errorf("expected ArchivedAt 2024-06-01T12:00:00Z, got %s", data.ArchivedAt.ValueString())
-	}
+const testAccVaultResourceBasicConfig = `
+resource "anthropic_vault" "test" {
+  display_name = "tf-acc-test-vault-basic"
 }
+`
 
-func TestMapVaultResponseToState_MetadataPopulated(t *testing.T) {
-	vault := &anthropic.BetaManagedAgentsVault{
-		ID:          "vlt_04JKL",
-		DisplayName: "vault-with-meta",
-		Type:        anthropic.BetaManagedAgentsVaultTypeVault,
-		CreatedAt:   time.Now(),
-		UpdatedAt:   time.Now(),
-		Metadata: map[string]string{
-			"team": "platform",
-			"env":  "prod",
+const testAccVaultResourceWithMetadataConfig = `
+resource "anthropic_vault" "test" {
+  display_name = "tf-acc-test-vault-metadata"
+
+  metadata = {
+    team = "terraform"
+    env  = "test"
+  }
+}
+`
+
+func TestAccVaultResource_basic(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckVaultDestroyed,
+		Steps: []resource.TestStep{
+			// Create and Read
+			{
+				Config: testAccVaultResourceBasicConfig,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet("anthropic_vault.test", "id"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "display_name", "tf-acc-test-vault-basic"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "type", "vault"),
+					resource.TestCheckResourceAttrSet("anthropic_vault.test", "created_at"),
+					resource.TestCheckResourceAttrSet("anthropic_vault.test", "updated_at"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "archive_on_destroy", "false"),
+				),
+			},
+			// ImportState
+			{
+				ResourceName:            "anthropic_vault.test",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"archive_on_destroy"},
+			},
 		},
-	}
-
-	var data VaultResourceModel
-	diags := mapVaultResponseToState(vault, &data)
-	if diags.HasError() {
-		t.Fatalf("unexpected diagnostics: %v", diags)
-	}
-
-	if data.Metadata.IsNull() {
-		t.Fatal("expected Metadata to be non-null")
-	}
-
-	elements := data.Metadata.Elements()
-	if len(elements) != 2 {
-		t.Fatalf("expected 2 metadata entries, got %d", len(elements))
-	}
-
-	teamVal, ok := elements["team"]
-	if !ok {
-		t.Fatal("expected metadata key 'team'")
-	}
-	if teamVal.(types.String).ValueString() != "platform" {
-		t.Errorf("expected metadata team=platform, got %s", teamVal.(types.String).ValueString())
-	}
+	})
 }
 
-func TestMapVaultResponseToState_MetadataEmpty(t *testing.T) {
-	vault := &anthropic.BetaManagedAgentsVault{
-		ID:          "vlt_05MNO",
-		DisplayName: "vault-no-meta",
-		Type:        anthropic.BetaManagedAgentsVaultTypeVault,
-		CreatedAt:   time.Now(),
-		UpdatedAt:   time.Now(),
-		Metadata:    nil,
-	}
+func TestAccVaultResource_withMetadata(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckVaultDestroyed,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccVaultResourceWithMetadataConfig,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet("anthropic_vault.test", "id"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "display_name", "tf-acc-test-vault-metadata"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "metadata.team", "terraform"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "metadata.env", "test"),
+				),
+			},
+			{
+				ResourceName:            "anthropic_vault.test",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"archive_on_destroy"},
+			},
+		},
+	})
+}
 
-	var data VaultResourceModel
-	diags := mapVaultResponseToState(vault, &data)
-	if diags.HasError() {
-		t.Fatalf("unexpected diagnostics: %v", diags)
-	}
+const testAccVaultResourceUpdateConfigV1 = `
+resource "anthropic_vault" "test" {
+  display_name = "tf-acc-test-vault-update-v1"
 
-	if !data.Metadata.IsNull() {
-		t.Errorf("expected Metadata to be null when API returns nil, got %v", data.Metadata)
-	}
+  metadata = {
+    team = "terraform"
+    env  = "test"
+  }
+}
+`
+
+// V2 renames the vault, changes one metadata value, and removes another key —
+// exercising the PATCH-with-null clear path in buildMetadataPatch.
+const testAccVaultResourceUpdateConfigV2 = `
+resource "anthropic_vault" "test" {
+  display_name = "tf-acc-test-vault-update-v2"
+
+  metadata = {
+    team = "platform"
+  }
+}
+`
+
+func TestAccVaultResource_update(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckVaultDestroyed,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccVaultResourceUpdateConfigV1,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("anthropic_vault.test", "display_name", "tf-acc-test-vault-update-v1"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "metadata.team", "terraform"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "metadata.env", "test"),
+				),
+			},
+			{
+				Config: testAccVaultResourceUpdateConfigV2,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("anthropic_vault.test", "display_name", "tf-acc-test-vault-update-v2"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "metadata.team", "platform"),
+					// env was removed in config and must be cleared server-side.
+					resource.TestCheckNoResourceAttr("anthropic_vault.test", "metadata.env"),
+				),
+			},
+		},
+	})
+}
+
+const testAccVaultResourceArchiveOnDestroyConfig = `
+resource "anthropic_vault" "test" {
+  display_name       = "tf-acc-test-vault-archive-on-destroy"
+  archive_on_destroy = true
+}
+`
+
+func TestAccVaultResource_archiveOnDestroy(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckVaultArchivedAndCleanup,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccVaultResourceArchiveOnDestroyConfig,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet("anthropic_vault.test", "id"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "display_name", "tf-acc-test-vault-archive-on-destroy"),
+					resource.TestCheckResourceAttr("anthropic_vault.test", "archive_on_destroy", "true"),
+				),
+			},
+			// ImportState — archive_on_destroy is local-only; provider defaults it to false on import.
+			{
+				ResourceName:            "anthropic_vault.test",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"archive_on_destroy"},
+			},
+		},
+	})
 }
